Split operator classes into separate files
Operators keep growing in size and the file length has
already surpassed 1000 lines. Figured it was time
to separate them for better readability.
